Improves TechnicalData Utils
